Revert "drm: Use size_t type for len variable in drm_copy_field()"
This reverts commit df5ac9392648394db23a98442d58f283b5479329.

+++ drivers/gpu/drm/drm_ioctl.c
@@ -472,7 +472,7 @@ EXPORT_SYMBOL(drm_invalid_op);
  */
 static int drm_copy_field(char __user *buf, size_t *buf_len, const char *value)
 {
-	size_t len;
+	int len;
 
 	/* don't overflow userbuf */
 	len = strlen(value);
